Move client Suspense boundaries to after usePreloadedQuery
It seems that using `usePreloadedQuery` inside a `<Suspense>` on a client component will cause it to always suspend on the client, which it doesn't need to. If it does, an hydration error occurs.

Moving to after `usePreloadedQuery`, to the data or fragments that it renders works fine however.

We still need it to be able to render a fallback in case a fragment is refetched.